brief: the family crosses — the document, its reader, and the delegation trio
Five routes under one prefix: the sweep-if-due document read with its absent
literal and the fold, the line verdict with its three 404 sentences, the
monotonic read cursor, the delegation pair (a grant fires the parked drafts),
and the draft verdict — the one route that causes something to leave, so the
one that refuses on staleness (409, the exact sentence). The engine's reader
plane crossed with them; the writer was already there from the batch-4 job.

The 39-case byte-diff runs warm (one document per person per day — mutations
work per-side probe rows and seeded resolved pairs) and caught four decode
bugs on its first run: batch/supersedes uuid → ::text; the seq claim,
mark_brief_stale, and the mirror's read_seq INT4 → ::int8. The 'pending'
absent kind is unreachable at the harness's hours — pinned at the engine
level, recorded in the migration doc.

+++ api/src/daily_brief/artifact.rs
@@ -59,7 +59,7 @@ pub async fn mirror_brief_artifact(
 ) -> Result<(), sqlx::Error> {
     let row: Option<MirrorRow> = sqlx::query_as(
         "select to_char(b.brief_date, 'YYYY-MM-DD'), b.artifact_id::text, a.folder_id::text, \
-                b.read_seq, b.agent_name, u.email, u.name \
+                b.read_seq::int8, b.agent_name, u.email, u.name \
          from daily_briefs b join users u on u.id = b.user_id \
          left join artifacts a on a.id = b.artifact_id \
          where b.id = $1::uuid and b.user_id = $2::uuid",
